chore: move proof info to wallet
