+++ crates/cdk-common/src/common.rs
@@ -1,15 +1,13 @@
 //! Types
 
 use serde::{Deserialize, Serialize};
-use uuid::Uuid;
 
 use crate::error::Error;
-use crate::mint_url::MintUrl;
 use crate::nuts::nut00::ProofsMethods;
-use crate::nuts::{
-    CurrencyUnit, MeltQuoteState, PaymentMethod, Proof, Proofs, PublicKey, SpendingConditions,
-    State,
-};
+use crate::nuts::{CurrencyUnit, MeltQuoteState, PaymentMethod, Proofs};
+// Re-export ProofInfo from wallet module for backwards compatibility
+#[cfg(feature = "wallet")]
+pub use crate::wallet::ProofInfo;
 use crate::Amount;
 
 /// Result of a finalized melt operation
@@ -156,123 +154,6 @@ impl std::fmt::Debug for FinalizedMelt {
     }
 }
 
-/// Prooinfo
-#[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ProofInfo {
-    /// Proof
-    pub proof: Proof,
-    /// y
-    pub y: PublicKey,
-    /// Mint Url
-    pub mint_url: MintUrl,
-    /// Proof State
-    pub state: State,
-    /// Proof Spending Conditions
-    pub spending_condition: Option<SpendingConditions>,
-    /// Unit
-    pub unit: CurrencyUnit,
-    /// Operation ID that is using/spending this proof
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub used_by_operation: Option<Uuid>,
-    /// Operation ID that created this proof
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub created_by_operation: Option<Uuid>,
-}
-
-impl ProofInfo {
-    /// Create new [`ProofInfo`]
-    pub fn new(
-        proof: Proof,
-        mint_url: MintUrl,
-        state: State,
-        unit: CurrencyUnit,
-    ) -> Result<Self, Error> {
-        let y = proof.y()?;
-
-        let spending_condition: Option<SpendingConditions> = (&proof.secret).try_into().ok();
-
-        Ok(Self {
-            proof,
-            y,
-            mint_url,
-            state,
-            spending_condition,
-            unit,
-            used_by_operation: None,
-            created_by_operation: None,
-        })
-    }
-
-    /// Create new [`ProofInfo`] with operation tracking
-    pub fn new_with_operations(
-        proof: Proof,
-        mint_url: MintUrl,
-        state: State,
-        unit: CurrencyUnit,
-        used_by_operation: Option<Uuid>,
-        created_by_operation: Option<Uuid>,
-    ) -> Result<Self, Error> {
-        let y = proof.y()?;
-
-        let spending_condition: Option<SpendingConditions> = (&proof.secret).try_into().ok();
-
-        Ok(Self {
-            proof,
-            y,
-            mint_url,
-            state,
-            spending_condition,
-            unit,
-            used_by_operation,
-            created_by_operation,
-        })
-    }
-
-    /// Check if [`Proof`] matches conditions
-    pub fn matches_conditions(
-        &self,
-        mint_url: &Option<MintUrl>,
-        unit: &Option<CurrencyUnit>,
-        state: &Option<Vec<State>>,
-        spending_conditions: &Option<Vec<SpendingConditions>>,
-    ) -> bool {
-        if let Some(mint_url) = mint_url {
-            if mint_url.ne(&self.mint_url) {
-                return false;
-            }
-        }
-
-        if let Some(unit) = unit {
-            if unit.ne(&self.unit) {
-                return false;
-            }
-        }
-
-        if let Some(state) = state {
-            if !state.contains(&self.state) {
-                return false;
-            }
-        }
-
-        if let Some(spending_conditions) = spending_conditions {
-            match &self.spending_condition {
-                None => {
-                    if !spending_conditions.is_empty() {
-                        return false;
-                    }
-                }
-                Some(s) => {
-                    if !spending_conditions.contains(s) {
-                        return false;
-                    }
-                }
-            }
-        }
-
-        true
-    }
-}
-
 /// Key used in hashmap of ln backends to identify what unit and payment method
 /// it is for
 #[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize)]
@@ -319,11 +200,8 @@ impl Default for QuoteTTL {
 mod tests {
     use std::str::FromStr;
 
-    use cashu::SecretKey;
-
-    use super::{FinalizedMelt, ProofInfo};
-    use crate::mint_url::MintUrl;
-    use crate::nuts::{CurrencyUnit, Id, Proof, PublicKey, SpendingConditions, State};
+    use super::FinalizedMelt;
+    use crate::nuts::{Id, Proof, PublicKey};
     use crate::secret::Secret;
     use crate::Amount;
 
@@ -388,90 +266,6 @@ mod tests {
         assert_eq!(finalized.amount(), Amount::from(31));
         assert_eq!(finalized.fee_paid(), Amount::from(1));
         assert_eq!(finalized.total_amount(), Amount::from(32));
-    }
-
-    #[test]
-    fn test_matches_conditions() {
-        let keyset_id = Id::from_str("00deadbeef123456").unwrap();
-        let proof = Proof::new(
-            Amount::from(64),
-            keyset_id,
-            Secret::new("test_secret"),
-            PublicKey::from_hex(
-                "02deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef",
-            )
-            .unwrap(),
-        );
-
-        let mint_url = MintUrl::from_str("https://example.com").unwrap();
-        let proof_info =
-            ProofInfo::new(proof, mint_url.clone(), State::Unspent, CurrencyUnit::Sat).unwrap();
-
-        // Test matching mint_url
-        assert!(proof_info.matches_conditions(&Some(mint_url.clone()), &None, &None, &None));
-        assert!(!proof_info.matches_conditions(
-            &Some(MintUrl::from_str("https://different.com").unwrap()),
-            &None,
-            &None,
-            &None
-        ));
-
-        // Test matching unit
-        assert!(proof_info.matches_conditions(&None, &Some(CurrencyUnit::Sat), &None, &None));
-        assert!(!proof_info.matches_conditions(&None, &Some(CurrencyUnit::Msat), &None, &None));
-
-        // Test matching state
-        assert!(proof_info.matches_conditions(&None, &None, &Some(vec![State::Unspent]), &None));
-        assert!(proof_info.matches_conditions(
-            &None,
-            &None,
-            &Some(vec![State::Unspent, State::Spent]),
-            &None
-        ));
-        assert!(!proof_info.matches_conditions(&None, &None, &Some(vec![State::Spent]), &None));
-
-        // Test with no conditions (should match)
-        assert!(proof_info.matches_conditions(&None, &None, &None, &None));
-
-        // Test with multiple conditions
-        assert!(proof_info.matches_conditions(
-            &Some(mint_url),
-            &Some(CurrencyUnit::Sat),
-            &Some(vec![State::Unspent]),
-            &None
-        ));
-    }
-
-    #[test]
-    fn test_matches_conditions_with_spending_conditions() {
-        // This test would need to be expanded with actual SpendingConditions
-        // implementation, but we can test the basic case where no spending
-        // conditions are present
-
-        let keyset_id = Id::from_str("00deadbeef123456").unwrap();
-        let proof = Proof::new(
-            Amount::from(64),
-            keyset_id,
-            Secret::new("test_secret"),
-            PublicKey::from_hex(
-                "02deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef",
-            )
-            .unwrap(),
-        );
-
-        let mint_url = MintUrl::from_str("https://example.com").unwrap();
-        let proof_info =
-            ProofInfo::new(proof, mint_url, State::Unspent, CurrencyUnit::Sat).unwrap();
-
-        // Test with empty spending conditions (should match when proof has none)
-        assert!(proof_info.matches_conditions(&None, &None, &None, &Some(vec![])));
-
-        // Test with non-empty spending conditions (should not match when proof has none)
-        let dummy_condition = SpendingConditions::P2PKConditions {
-            data: SecretKey::generate().public_key(),
-            conditions: None,
-        };
-        assert!(!proof_info.matches_conditions(&None, &None, &None, &Some(vec![dummy_condition])));
     }
 }
