+++ crates/cdk/src/wallet/swap/saga/mod.rs
@@ -28,7 +28,7 @@ use crate::amount::SplitTarget;
 use crate::dhke::construct_proofs;
 use crate::nuts::nut00::ProofsMethods;
 use crate::nuts::{nut10, Proofs, SpendingConditions, State};
-use crate::types::ProofInfo;
+use cdk_common::wallet::ProofInfo;
 use crate::wallet::saga::{
     add_compensation, clear_compensations, execute_compensations, new_compensations, Compensations,
     RevertProofReservation as RevertSwapProofReservation,
